Add a privacy policy page for me. I will send the sample that my friend used for his school project. I want you to be creative with my privacy policy:
Privacy policy
Last Revised: February 5th 2017

Introduction

NovaMergeXcap Limited (in incorporation) (“us“, “we” or “Company“) respect the privacy of our users (each, “you” or “User“) and are committed to protect the privacy of Users who access, download, install or register to our mobile application (the “Application“), our website or any other online services we provide (collectively: the “Services“). The Company has papered this Privacy Policy to outline our practices with respect to collecting, using and disclosing your information when you use the Services. We encourage you to read the Privacy Policy carefully and use it to make informed decisions. By using the Services, you agree to the terms of this Privacy Policy and your continued use of the Services constitutes your ongoing agreement to the Privacy Policy. The Privacy Policy is a part of the Terms of Service and is incorporated there in by reference. In this Privacy Policy you will read about, among other things:

* What type of information we collect

* Cookies and Google Analytics

* How we use the information

* With whom we share the information and for what purpose

* For how long we retain the information

* How we protect your information

* Advertisements

* Advertising ID and Advertising Identifier

* How to contact us