I tried to sign up a new user and this was my experience:
I successfully signed up a user and the email reflected in my firebase console.

During the onboarding process where I collect the account type of the user, phone number, etc. After I click the save and continue button, I get this error message pop up:
Update failed
internal server error during registration

and I also do not get navigated to the dashboard. I think there is a flaw in the sign in process and this is usually the easy part of back end set up, the authentication so I am really wondering if you are up for the task